% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiCopSelectMix.r
\name{BiCopSelectMix}
\alias{BiCopSelectMix}
\title{Bivariate Copula Selection with Mixed continuous and discrete variables}
\usage{
BiCopSelectMix(
  u1Plus,
  u1Minus,
  u2Plus,
  u2Minus,
  isDisc1,
  isDisc2,
  familyset,
  selectioncrit = "AIC",
  max.df = 30,
  max.BB = list(BB1 = c(5, 6), BB6 = c(6, 6), BB7 = c(5, 6), BB8 = c(6, 1)),
  debug = FALSE
)
}
\arguments{
\item{u1Plus}{uscore1+ or cdf at right limit.}

\item{u1Minus}{uscore1- or cdf at left limit.}

\item{u2Plus}{uscore2+ or cdf at right limit.}

\item{u2Minus}{uscore2- or cdf at left limit.}

\item{isDisc1}{Boolean. TRUE if the first variable is discrete.}

\item{isDisc2}{Boolean. TRUE if the second variable is discrete.}

\item{familyset}{Vector of bivariate copula families to select from.
The vector has to include at least one bivariate copula
family that allows for positive and one that allows for negative dependence.
If \code{familyset = NA} (default), selection among all possible families is
performed. If a vector of negative numbers is provided, selection among all
but \code{abs(familyset)} families is performed. Coding of bivariate copula
families: \cr
\code{0} = independence copula \cr
\code{1} = Gaussian copula \cr
\code{2} = Student t copula (t-copula) \cr
\code{3} = Clayton copula \cr
\code{4} = Gumbel copula \cr
\code{5} = Frank copula \cr
\code{6} = Joe copula \cr
\code{7} = BB1 copula \cr
\code{8} = BB6 copula \cr
\code{9} = BB7 copula \cr
\code{10} = BB8 copula \cr
\code{13} = rotated Clayton copula (180 degrees; ``survival Clayton'') \cr
\code{14} = rotated Gumbel copula (180 degrees; ``survival Gumbel'') \cr
\code{16} = rotated Joe copula (180 degrees; ``survival Joe'') \cr
\code{17} = rotated BB1 copula (180 degrees; ``survival BB1'')\cr
\code{18} = rotated BB6 copula (180 degrees; ``survival BB6'')\cr
\code{19} = rotated BB7 copula (180 degrees; ``survival BB7'')\cr
\code{20} = rotated BB8 copula (180 degrees; ``survival BB8'')\cr
\code{23} = rotated Clayton copula (90 degrees) \cr
\code{24} = rotated Gumbel copula (90 degrees) \cr
\code{26} = rotated Joe copula (90 degrees) \cr
\code{27} = rotated BB1 copula (90 degrees) \cr
\code{28} = rotated BB6 copula (90 degrees) \cr
\code{29} = rotated BB7 copula (90 degrees) \cr
\code{30} = rotated BB8 copula (90 degrees) \cr
\code{33} = rotated Clayton copula (270 degrees) \cr
\code{34} = rotated Gumbel copula (270 degrees) \cr
\code{36} = rotated Joe copula (270 degrees) \cr
\code{37} = rotated BB1 copula (270 degrees) \cr
\code{38} = rotated BB6 copula (270 degrees) \cr
\code{39} = rotated BB7 copula (270 degrees) \cr
\code{40} = rotated BB8 copula (270 degrees) \cr
\code{104} = Tawn type 1 copula \cr
\code{114} = rotated Tawn type 1 copula (180 degrees) \cr
\code{124} = rotated Tawn type 1 copula (90 degrees) \cr
\code{134} = rotated Tawn type 1 copula (270 degrees) \cr
\code{204} = Tawn type 2 copula \cr
\code{214} = rotated Tawn type 2 copula (180 degrees) \cr
\code{224} = rotated Tawn type 2 copula (90 degrees) \cr
\code{234} = rotated Tawn type 2 copula (270 degrees) \cr}

\item{selectioncrit}{Character indicating the criterion for bivariate copula
selection. Possible choices: \code{selectioncrit = "AIC"} (default),
\code{"BIC"}, or \code{"logLik"}.}

\item{max.df}{Numeric; upper bound for the estimation of the degrees of
freedom parameter of the t-copula (default: \code{max.df = 30}).}

\item{max.BB}{List; upper bounds for the estimation of the two parameters
(in absolute values) of the BB1, BB6, BB7 and BB8 copulas \cr (default:
\code{max.BB = list(BB1=c(5,6),BB6=c(6,6),BB7=c(5,6),BB8=c(6,1))}).}

\item{debug}{Debug flag; default is FALSE}
}
\value{
A \code{BiCop} object.
}
\description{
Selection and Maximum Likelihood Estimation of Bivariate Copula Families with
Continuous-Discrete Mixed Variables.
}
\details{
See \link[VineCopula]{BiCopSelect}.
Part of the code is adapted from \link[VineCopula]{BiCopSelect}.
Need to be careful what rotation means if bivariate copula is not
permutation symmetric. \cr
\eqn{(U_1,U_2)} positively dependent implies the copulas of
\eqn{(1-U_1,U_2)} and \eqn{(U_1,1-U_2)} have negative dependence. 
Refer to these as 1-reflected, 2-reflected;
and \eqn{(1-U_1,1-U_2)} as reflected or (12)-reflected.\cr
Counterclockwise rotation \eqn{(U_1,U_2)} in the unit square about the centre (1/2,1/2) leads to the following.\cr
Rotation of \eqn{\pi/2}:   \eqn{(U_1,U_2)} to \eqn{(1-U_2,U_1)}.\cr
Rotation of \eqn{2\pi/2}:  \eqn{(U_1,U_2)} to \eqn{(1-U_1,1-U_2)}.\cr
Rotation of \eqn{3\pi/2}:  \eqn{(U_1,U_2)} to \eqn{(U_2,1-U_1)}.\cr
Check if rotation in VineCopula is clockwise or counterclockwise and update here.
}
