% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisect_extra.r
\name{bisect_extra}
\alias{bisect_extra}
\title{Bisection method for finding a unique root}
\usage{
bisect_extra(fun, x0, x1, tol = 1e-05, niter = 18, debug = FALSE, plus, minus)
}
\arguments{
\item{fun}{vector of strictly increasing functions on [0,1] which cross 0}

\item{x0}{vector of values where fun<0}

\item{x1}{vector of values where fun>0}

\item{tol}{tolerance on consecutive iterates for stopping, default is 1.e-05}

\item{niter}{maximum number of iterations, default is 18}

\item{debug}{debug flag for intermediate steps, default is FALSE}
}
\value{
vector of roots
}
\description{
Bisection method for a function which is monotone increasing in \eqn{[0,1]}.
Find root fun(v)=0 where fun is a function that returns a vector.
}
\examples{
out <- bisect(function(x) sin(x)-c(.25,.5), c(0,0), c(1,1), niter=20, debug=TRUE)
}
