% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_mat.r
\name{cor_mat}
\alias{cor_mat}
\title{correlation matrix : vander Waarden/ polychoric/ polyserial}
\usage{
cor_mat(data, is_disc = rep(FALSE, ncol(data)), debug = FALSE)
}
\arguments{
\item{data}{dataframe or matrix}

\item{is_disc}{logical vector, TRUE for the i-th variable if it is discrete. By default all FALSE.}

\item{debug}{debug flag}
}
\value{
correlation matrix
}
\description{
A correlation matrix of mixed continuous/discrete data.
For continuous-continuous, calculate the vander Waarden correlation of the normal scores.
For continuous-discrete, calculate the polyserial correlation using \code{polycor::polychor}.
For discrete-discrete, calculate the polychoric correlation using \code{polychor::polyserial}.
}
\examples{
x = matrix(rnorm(700), nrow = 100, ncol = 7)
x[, c(1,2,3)] = round(x[, c(1,2,3)])
cor_mat(x, c(TRUE, TRUE, TRUE, FALSE, FALSE, FALSE, FALSE))

}
