% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbvn.r
\name{dbvn}
\alias{dbvn}
\title{bivariate normal density (internal function)}
\usage{
dbvn(x, rho, ilog = FALSE)
}
\arguments{
\item{x}{vector of length 2 or 2-column matrix}

\item{rho}{correlation parameter with -1 < rho < 1}

\item{ilog}{TRUE for log density, FALSE for density}
}
\value{
bivariate standard normal density with correlation rho, or log density
}
\description{
bivariate normal density with zero means and unit variances
}
\examples{
rho = .8
N = 1000
x = qnorm(VineCopula::BiCopSim(N, 1, rho, par2 = 0, obj = NULL, check.pars = TRUE))
dbvn(x, rho)
}
