% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varray-CopulaModel.r
\name{Dvinearray}
\alias{Dvinearray}
\title{Vine array for D-vine}
\usage{
Dvinearray(d, iNO = FALSE)
}
\arguments{
\item{d}{dimension}

\item{iNO}{FALSE for standard form (default) or TRUE for natural order}
}
\value{
vine array for D-vine
}
\description{
Vine array for D-vine in standard form or normal order (latter is only useful for enumerating vines)
}
\examples{
Dvinearray(6) 
Dvinearray(5) # same as 5x5 submatrix of above

}
