% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussMST-cfi.r
\name{edges2array.trunc}
\alias{edges2array.trunc}
\title{Edge matrix of a truncated vine to a vine array}
\usage{
edges2array.trunc(d, mtree, edgemat, pcv, iprint = FALSE)
}
\arguments{
\item{d}{dimension or number of variables}

\item{mtree}{truncation level}

\item{edgemat}{edge matrix of dimension [(d-1)+...+(d-mtree)] x (mtree+3)
  with columns node1, node2, vine tree level, given[1], ..., given[mtree]
(it is not checked if this matches a proper vine array).
Each row represents an edge of the vine.}

\item{pcv}{corresponding vector of partial correlations for edges of truncated vine}

\item{iprint}{print flag for intermediate results, default FALSE}
}
\value{
Returns a list with the following named components:
\enumerate{
 \item \code{VineA} -  truncated vine array of dimension dxd, with 0s for trees at level greater than mtree
 \item \code{pcmat} -  matrix of partial correlations by tree
}
}
\description{
Convert a truncated vine stored as an edge matrix (columns node1, node2, level, conditioning variables) 
to a truncated vine array
}
\details{
WARNING: It has not been checked what happens with incorrect input.
}
\examples{
edgemat6=matrix(c(
   1,2, 1, 0,0,0,0,
   1,3, 1, 0,0,0,0,
   2,4, 1, 0,0,0,0,
   1,5, 1, 0,0,0,0,
   2,6, 1, 0,0,0,0,
   2,3, 2, 1,0,0,0,
   1,4, 2, 2,0,0,0,
   3,5, 2, 1,0,0,0,
   4,6, 2, 2,0,0,0,
   3,4, 3, 2,1,0,0,
   2,5, 3, 1,3,0,0,
   1,6, 3, 2,4,0,0,
   4,5, 4, 1,3,2,0,
   3,6, 4, 2,4,1,0,
   5,6, 5, 2,4,1,3), 15,7, byrow=TRUE)
colnames(edgemat6)=c("node1","node2","level","giv1","giv2","giv3","giv4")
edgemat.trunc2=edgemat6[1:9,1:5]
pcvec2=c(.6,.7,.6,.7,.5, .4,.4,.3,.3) 
edges2array.trunc(d=6,mtree=2,edgemat.trunc2,pcvec2)
edgemat.trunc3=edgemat6[1:12,1:6]
pcvec3=c(.6,.7,.6,.7,.5, .4,.4,.3,.3, .2,.2,.1) 
edges2array.trunc(d=6,mtree=3,edgemat.trunc3,pcvec3)
iperm=c(2,4,12,10,9,3,8,5,11,1,6,7)
edges2array.trunc(d=6,mtree=3,edgemat.trunc3[iperm,],pcvec3[iperm])
# should be the same, because row permutation does not affect vine

}
