% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cfi.r
\name{get_cfi_from_pc}
\alias{get_cfi_from_pc}
\title{Get CFI for a partial correlation array based on a truncation level}
\usage{
get_cfi_from_pc(rmat, pc, n, trunc_level)
}
\arguments{
\item{rmat}{Correlation matrix.}

\item{pc}{Partial correlation array (computed using get_cfi_from_vine_array)}

\item{n}{Sample size.}

\item{trunc_level}{Truncation level.}
}
\value{
CFI
}
\description{
A helper function that computes the comparative fit index (CFI)
for a given correlation matrix and partial correlation array and truncation level
}
\details{
CFI (comparative fit index) is defined as follows.
CFI = 1- numerator/denominator,\cr
numerator = \eqn{\max(0,D_t-\nu_t)},\cr
denominator = \eqn{\max(0,D_t-\nu_t,D_0-\nu_0)},\cr
\eqn{D_0=-n\log(\det(R))},\cr
\eqn{D_t = n[-L_t(V) - \log(\det(R))]} is decreasing as t increases,\cr
\eqn{L_t(V) = -\sum_{1:t} \sum_e -\log(1-r_e^2)} (sum pcor up to tree t),\cr
\eqn{\nu_t = (d-t)(d-t-1)/2; \nu_0=d(d-1)/2},\cr
\eqn{L_t(V)} is increasing in t.
}
