% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cfi.r
\name{get_cfi_from_vine_array}
\alias{get_cfi_from_vine_array}
\title{Get CFI from a vine array based on a truncation level}
\usage{
get_cfi_from_vine_array(rmat, vine_array, n, trunc_level)
}
\arguments{
\item{rmat}{Correlation matrix.}

\item{vine_array}{Vine array.}

\item{n}{Sample size.}

\item{trunc_level}{Truncation level.}
}
\value{
CFI
}
\description{
A helper function that computes the comparative fit index (CFI)
for a given correlation matrix and vine array and truncation level
}
