% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vine_array.r
\name{helper_append_y}
\alias{helper_append_y}
\title{Appended vine array with response variable y}
\usage{
helper_append_y(rmat, vine_array_x, trunc_level = nrow(rmat) - 1)
}
\arguments{
\item{rmat}{Correlation matrix of covariates and response; vander Waarden/ polychoric/ polyserial correlations.}

\item{vine_array_x}{Vine array of the explanatory variables.}

\item{trunc_level}{Truncation level}
}
\value{
Full vine array (one more row/column compared with input); the response variable is in the last column.
}
\description{
Given a vine array for covariates, this helper function adds the response variable so that max partial correlations are achieved subject to the proximity condition for vines.
}
