% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_mat.r
\name{isposdef}
\alias{isposdef}
\title{Is positive definite?}
\usage{
isposdef(amat)
}
\arguments{
\item{amat}{symmetric matrix}
}
\value{
TRUE if amat is positive definite, FALSE otherwise
}
\description{
Check if a square symmetric matrix is positive definite.
}
\examples{
a1=matrix(c(1,.5,.5,1),2,2); print(isposdef(a1))
a2=matrix(c(1,1.5,1.5,1),2,2); print(isposdef(a2))

}
