% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nscore.r
\name{nscore}
\alias{nscore}
\title{Normal scores}
\usage{
nscore(data, iopt = FALSE)
}
\arguments{
\item{data}{dataframe or matrix, or vector, of reals}

\item{iopt}{flag with default FALSE in which case adjustment -1/2, set as TRUE to use adjustment 'a' in nscoreOpta()}
}
\value{
matrix or vector of normal scores
}
\description{
Transform each variable to normal scores
}
\details{
For a variable y (column of data) with sample size n, first converted to rank, then
to zscore=(rank+adj)/(n+1+2*adj) where default adj=-0.5, or
adj=optimaladjustment so that mean of  \eqn{zscore^2} is 1.
}
\examples{
set.seed(123)
x=matrix(rexp(400),200,2)
summary(x)
z=nscore(x)
summary(z) # means close to 0
cov(z)  # variances close to 1

}
