% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_parameter2.r
\name{parse_parameter2}
\alias{parse_parameter2}
\title{parse bivariate copula parameters for likelihood estimation (internal function)}
\usage{
parse_parameter2(paramvec, fam_vec, d)
}
\arguments{
\item{paramvec}{vector of copula parameters to be parsed, the first d-1 elements are the first parameters for the copula models, the remaining elements (if any) are the second parameters for the copula models}

\item{fam_vec}{vector of copula families of length d-1}

\item{d}{dimension of the dataset}
}
\value{
the parsed parameter vector
}
\description{
Parse bivariate copula parameters for likelihood estimation so that all parameters are within the range. Note that parameters for Gaussian and t copulas before parsing are in the range of (-infinity, infinity), parameters for Gaussian and t copulas after parsing are in the range of (-1, 1). A tanh(x/10) transformation is taken.
}
