% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialcor-CopulaModel.r
\name{partcor}
\alias{partcor}
\title{Partial correlation}
\usage{
partcor(S, given, j, k)
}
\arguments{
\item{S}{covariance or correlation matrix, of dimension dxd}

\item{given}{vector indices for the given or conditioning variables}

\item{j, k}{indices for the conditioned variables}
}
\value{
partial correlation of variables j,k given indices in 'given'
}
\description{
Computation of a partial correlation from a covariance matrix
}
\details{
WARNING: It is not checked that j,k,given are indices in 1:d, where
  d is the dimension. It is not checked that intersect(c(j,k),given) is empty
}
\examples{
rr=toeplitz(c(1,.5,.25,.125,.05))
partcor(rr,c(1),3,4)
partcor(rr,c(1),3,5)
partcor(rr,c(1,2),3,4)
partcor(rr,c(1,2),3,5)

}
