% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlusToMinus.r
\name{PlusToMinus}
\alias{PlusToMinus}
\title{Plus to Minus for discrete variables}
\usage{
PlusToMinus(dataDisc, listSupport = NULL)
}
\arguments{
\item{dataDisc}{A discrete Uscore matrix.}

\item{listSupport}{A list of sorted support points, that is the possible values 
each variable can take. If not input, the function will produce it.}
}
\value{
Returns a list with the following named components:
\enumerate{
\item \code{dataMinus} - A discrete Uscore (cdf) matrix wih the same dimension as dataDisc
\item \code{listSupport} - List of support points for each discrete variable
}
}
\description{
Given a discrete Uscore (or cdf) matrix F^+_i = P(X_i <= x_i), 
calculate the left limit of the cdf F^-_i = P(X_i < x_i).
Internal function.
}
