% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussMST-cfi.r
\name{proxcheck}
\alias{proxcheck}
\title{Proximity condition check}
\usage{
proxcheck(avec, bvec)
}
\arguments{
\item{avec}{(with m distinct values) sorted in increasing order}

\item{bvec}{(with m distinct values) sorted in increasing order}
}
\value{
Returns a list with the following named components:
\enumerate{
 \item \code{iok} - equals 1 if proximity condition is satisfied, 0 if not
 \item \code{icomv} - vector of common variables (length m-1 if iok=1)
 \item \code{icond1} - smaller index of two conditioned variables if iok=1
 \item \code{icond2} - larger index of two conditioned variables if iok=1
}
}
\description{
Check of proximity condition for ordered vectors avec, bvec of length m
}
\examples{
proxcheck(c(1,4,5),c(2,4,6))
proxcheck(c(1,4,5),c(1,4,6))

}
