% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvine_cop_select_mix.r
\name{rvine_cop_select_mix}
\alias{rvine_cop_select_mix}
\title{R-vine copula selection for mixed continuous/discrete variables}
\usage{
rvine_cop_select_mix(
  data,
  is_disc,
  familyset = NA,
  vine_array,
  selectioncrit = "AIC",
  trunc_level = NA,
  cores = 1,
  debug = FALSE
)
}
\arguments{
\item{data}{An \eqn{n}-by-\eqn{d} data matrix with uniform margins.
The last column corresponds the response variable
and the first \eqn{d-1} columns correspond to the explanatory variables.}

\item{is_disc}{A boolean vector of length \eqn{d} indicating
whether each variable is discrete.}

\item{familyset}{A vector of bivariate copula families to select from;
this uses the family codes in VineCopula.}

\item{vine_array}{A \eqn{d}-by-\eqn{d} matrix representing the vine array.
The last column corresponds to the response variable.}

\item{selectioncrit}{The criterion for pair-copula selection.
Possible choices: selectioncrit = "AIC" (default), "BIC", or "logLik".}

\item{trunc_level}{The level of truncation.}

\item{cores}{integer; if \code{cores > 1}, estimation will be parallelized
within each tree (using \code{\link[foreach]{foreach}}).}

\item{debug}{Debug flag; default is FALSE}
}
\value{
A \code{VineCopula::RVineMatrix} object
}
\description{
Bivariate copula selection for mixed continuous/discrete variables
}
\details{
Adapted from \code{VineCopula::RVineCopSelect}
}
