% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvinesimvec_disc.r
\name{rvinesimvec_disc}
\alias{rvinesimvec_disc}
\title{simulation for R-vine copulas with mixed discrete/continuous variables}
\usage{
rvinesimvec_disc(
  nsim,
  A,
  fam,
  param1,
  param2,
  is_disc,
  discrete_values = c(-1.5, -0.5, 0.5, 1.5),
  marginal_dist = "pnorm",
  marginal_param = c(0, 1),
  iprint = F
)
}
\arguments{
\item{nsim}{sample size for simulation}

\item{A}{d*d vine array, or ntrunc*d vine array as only ntrunc rows are used}

\item{fam}{d by d matrix for the copula families coded by the VineCopula package}

\item{param1}{d by d matrix for the first parameter of the copula models}

\item{param2}{d by d matrix for the second parameter of the copula models}

\item{is_disc}{vector of length d representing if the variable is discrete or not}

\item{discrete_values}{possible values for the discrete variables to take at the original scale}

\item{marginal_dist}{string of the pdf function of the marginal distribution (for discrete variables)}

\item{marginal_param}{vector of the parameters of the marginal distribution (for discrete variables)}

\item{iprint}{print flag for intermediate results}
}
\value{
the simulated u-scores
}
\description{
simulating u-scores from a R-vine copulas with mixed discrete/continuous variables based on rosenblatt transformation
}
