% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_copula_structure_estimation.r
\name{survival_copula_structure_estimation}
\alias{survival_copula_structure_estimation}
\title{vine copula structure estimation with a survival response}
\usage{
survival_copula_structure_estimation(
  data_cdf,
  censor_status,
  is_disc,
  prlevel = 2,
  all_fam = c(1, 2, 4, 14, 5, 7, 10, 20, 17)
)
}
\arguments{
\item{data_cdf:}{n by d matrix for u-scores}

\item{censor_status:}{vector of length n representing if the response variable is censored or not}

\item{is_disc:}{vector of length d representing if the variable is discrete or not}

\item{prlevel:}{print level for nlm}

\item{all_fam:}{all bivariate copula families to be considered as coded in VineCopula package}
}
\value{
a list with the following attributes:
\enumerate{
 \item \code{best_param} - estimated copula parameters associated with the censored response variable given the best vine structure and family matrix found
 \item \code{param1_complete} - matrix of the first parameters associated with the copula modesl for the predictors 
 \item \code{param2_complete} - matrix of the second parameters associated with the copula modesl for the predictors 
 \item \code{fam_estimated} - family matrix for the copula models
 \item \code{A_mst} - vine structure matrix
}
}
\description{
this function estimates the vine structure (such that the response variable is contained in a leaf node at every level of the vine), selects the bivariate copula families, and estimates the parameters associated with the selected copulas for a survival response variable with mixed discrete/continuous explanatory variables
}
