% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_predict.r
\name{survival_predict}
\alias{survival_predict}
\title{predict u-score of a survival response with mixed discrete/continuous explanatory variables at a quantile}
\usage{
survival_predict(
  p,
  paramvec,
  param1_complete,
  param2_complete,
  fam,
  A,
  data_cdf_plus,
  data_cdf_minus,
  is_disc,
  iprint = F
)
}
\arguments{
\item{p}{vector of conditional CDFs to be predicted of length m}

\item{paramvec}{vector of length (d-1) if all the copula families associated with the survival response variable are one-parameter families; vector of length (d-1)+ otherwise}

\item{param1_complete}{d-1 by d-1 matrix for the first parameter of the copula models for the predictors}

\item{param2_complete}{d-1 by d-1 matrix for the second parameter of the copula models for the predictors}

\item{fam}{d by d matrix for the copula families coded by the VineCopula package}

\item{A}{d by d matrix for the copula structure}

\item{data_cdf_plus}{n by (d-1) matrix for CDF from the right}

\item{data_cdf_minus}{n by (d-1) matrix for CDF from the left}

\item{is_disc}{vector of length d representing if the variable is discrete or not}
}
\value{
the conditional u-scores (n by m matrix) of the response variable given the predictors at quantiles p
}
\description{
predict u-score of a survival response with mixed discrete/continuous explanatory variables at quantile p such that the conditional CDF of the survival response given mixed discrete/continuous explanatory variables is equal to p
}
