% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_qqplot.r
\name{survival_qqplot}
\alias{survival_qqplot}
\title{generate the Q-Q plot of the empirical CDF values versus the CDF values from a parametric model for a right-censored variable}
\usage{
survival_qqplot(y, censor_status, pfunc, ...)
}
\arguments{
\item{y}{n by 1 numeric vector of the values of the right-censored variable}

\item{censor_status}{n by 1 logical vector, ith position TRUE if udata[i,d] is right-censored}

\item{pfunc}{function to get the CDF from a parametric model, parameters of the parametric model are provided as extra arguments at the end of the function call}
}
\value{
the Q-Q plot of the empirical CDF values versus the CDF values from a parametric model for a right-censored variable
}
\description{
generate the Q-Q plot of the empirical CDF values (on the y-axis) versus the CDF values from a parametric model (on the x-axis) for a right-censored variable
}
