% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalconditionalcdf.r
\name{survivalconditionalcdf}
\alias{survivalconditionalcdf}
\title{conditional CDF of the survival response given mixed discrete/continuous explanatory variables}
\usage{
survivalconditionalcdf(
  u,
  paramvec,
  param1_complete,
  param2_complete,
  fam,
  A,
  data_cdf_plus,
  data_cdf_minus,
  is_disc
)
}
\arguments{
\item{paramvec}{vector of length (d-1) if all the copula families associated with the survival response variable are one-parameter families; vector of length (d-1)+ otherwise}

\item{param1_complete}{d-1 by d-1 matrix for the first parameter of the copula models for the predictors}

\item{param2_complete}{d-1 by d-1 matrix for the second parameter of the copula models for the predictors}

\item{fam}{d by d matrix for the copula families coded by the VineCopula package}

\item{A}{d by d matrix for the copula structure}

\item{data_cdf_plus}{n by (d-1) matrix for CDF from the right}

\item{data_cdf_minus}{n by (d-1) matrix for CDF from the left}

\item{is_disc}{vector of length d representing if the variable is discrete or not}

\item{u-score}{scalar or vector of the response variable of length n}
}
\value{
the conditional CDF of the response variable given the explanatory variables of length n
}
\description{
conditional CDF of the survival response given mixed discrete/continuous explanatory variables
}
