% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalnllk.r
\name{survivalnllk}
\alias{survivalnllk}
\title{negative log-likelihood of vine copula model for survival response}
\usage{
survivalnllk(
  paramvec,
  param1_complete,
  param2_complete,
  fam,
  A,
  data_cdf_plus,
  data_cdf_minus,
  censor_status,
  is_disc
)
}
\arguments{
\item{paramvec}{vector of length (d-1) if all the copula families associated with the survival response variable are one-parameter families; vector of length (d-1)+ otherwise}

\item{param1_complete}{d-1 by d-1 matrix for the first parameter of the copula models for the predictors}

\item{param2_complete}{d-1 by d-1 matrix for the second parameter of the copula models for the predictors}

\item{fam}{d by d matrix for the copula families coded by the VineCopula package}

\item{A}{d by d matrix for the copula structure}

\item{data_cdf_plus}{n by (d-1) matrix for CDF from the right}

\item{data_cdf_minus}{n by (d-1) matrix for CDF from the left}

\item{censor_status}{vector of length n representing if the response variable is censored or not for each observation}

\item{is_disc}{vector of length d representing if each variable is discrete or not}
}
\value{
the negative log-likelihood of the specified vine copula model for a survival response
}
\description{
negative log-likelihood of a vine copula model for a survival response (the last column of the data matrix) and mixed discrete/continuous explanatory variables (the other columns of the data matrix)
}
