% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cfi.r
\name{trvine.CFI}
\alias{trvine.CFI}
\title{truncated vine based on comparative fit index}
\usage{
trvine.CFI(rmat, A, n, CFIbd, iprint = FALSE)
}
\arguments{
\item{rmat}{positive definite dxd correlation matrix, with d>2}

\item{A}{dxd vine array (only upper triangle is used)}

\item{n}{sample size used to compute rmat}

\item{CFIbd}{lower bound of CFI (comparative fit index) to truncate, default 0.95, set as 1.01 for no truncation}

\item{iprint}{set as T for intermediate prints, default=F}
}
\value{
Returns a list with the following named components:
\enumerate{
 \item \code{VineA} -  vine array of dimension dxd
 \item \code{pcmat} -  matrix of partial correlations by tree
 \item \code{treeweight} -  vector of length d-1 with
                \eqn{\sum_e \log(1-\rho_e^2)} for trees 1,...d-1
 \item \code{edgemat} -  matrix with columns node1 node2 level vector-of-conditioning
 \item \code{fitval} - vector with cumsum(treeweight)/sum(treeweight)
 \item \code{CFIv} - vector of CFI values
 \item \code{ntrunc} - truncation level to reach CFIbd
}
}
\description{
Apply CFI (comparative fit index) calculations to a correlation matrix and a vine array to get a truncated vine
}
\examples{
rmat=matrix(c(
  1.000,  0.172, -0.062,  0.385,  0.499, 0.267,  0.578,
  0.172,  1.000, -0.047,  0.493,  0.602, 0.396,  0.600,
 -0.062, -0.062,  1.000, -0.120, -0.092, 0.070, -0.070,
  0.385,  0.493, -0.120,  1.000,  0.607, 0.557,  0.742,
  0.499,  0.602, -0.092,  0.607,  1.000, 0.672,  0.883,
  0.267,  0.396,  0.070,  0.557,  0.672, 1.000,  0.685,
  0.578,  0.600, -0.070,  0.742,  0.883, 0.685,  1.000), 7,7)
A7=matrix(c( 
  5, 5, 5, 5, 5, 4, 5,
  0, 4, 4, 4, 4, 5, 4,
  0, 0, 1, 1, 6, 6, 1,
  0, 0, 0, 6, 1, 1, 6,
  0, 0, 0, 0, 2, 2, 2,
  0, 0, 0, 0, 0, 3, 3,
  0, 0, 0, 0, 0, 0, 7), 7,7, byrow=TRUE)
trvine.CFI(rmat, A7, n=400, CFIbd=0.98)

}
