% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varray-CopulaModel.r
\name{varrayperm}
\alias{varrayperm}
\title{Vine array with permuted indices}
\usage{
varrayperm(A, perm)
}
\arguments{
\item{A}{dxd vine array}

\item{perm}{permutation of 1:d}
}
\value{
vine array after permutation of indices
}
\description{
Create vine array with permuted indices, useful because some
algorithms for vines assume ordered 1:d on the diagonal.
}
\examples{
D5= matrix(c(1,0,0,0,0, 1,2,0,0,0, 2,1,3,0,0, 3,2,1,4,0, 4,3,2,1,5), 5,5)
D5perm= matrix(c(5,0,0,0,0, 5,4,0,0,0, 4,5,3,0,0, 3,4,5,2,0, 2,3,4,5,1), 5,5)
varrayperm(D5,perm=c(5,4,3,2,1))

}
