% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vWcor_discrete.r
\name{vWcor_discrete}
\alias{vWcor_discrete}
\title{van der Waerden correlation when first variable is discrete and second variable is right-censored}
\usage{
vWcor_discrete(dat1, dat2, censor_status, iprint = 0, iterlim = 100)
}
\arguments{
\item{dat1}{variable 1 that is discrete, vector of length n}

\item{dat2}{variable 2 that is right-censored, vector of length n}

\item{censor_status}{binary vector of length n, indicator of censored , =0 for uncensored, =1 for right-censored}

\item{iprint}{print_level for nlm minimization}

\item{iterlim}{number of iterations for nlm minimization}
}
\value{
van der Waerden (normal score) correlation value in the interval (-1,1)
}
\description{
van der Waerden correlation or correlation of normal scores when first variable is discrete and second variable is right-censored
}
\examples{
rho = .8
N = 1000
censoring_rate = 0.5
u = VineCopula::BiCopSim(N, 1, rho, par2 = 0, obj = NULL, check.pars = TRUE)
auxiliary = pnorm(rnorm(N, -sqrt(2) * qnorm(censoring_rate)))
censor_status = auxiliary < u[, 2]
u[as.logical(censor_status), 2] = auxiliary[as.logical(censor_status)]
x = qnorm(u)
x[, 1] = round(x[, 1])
vWcor(x[,1], x[,2], censor_status)

}
