% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vWcormat.r
\name{vWcormat}
\alias{vWcormat}
\title{van der Waerden correlation matrix when variable in the last column can be right-censored}
\usage{
vWcormat(udata, is_disc, censor_status, iprint = FALSE)
}
\arguments{
\item{udata}{n by d matrix with values in (0,1)}

\item{is_disc}{logical vector of length d, TRUE in jth position for discrete variable}

\item{censor_status}{n by 1 logical vector, ith position TRUE if udata[i,d] is right-censored}

\item{iprint}{flag for printing intermediate results if TRUE}
}
\value{
d by d positive definite van der Waerden (normal score) correlation matrix
}
\description{
van der Waerden correlation or correlation of normal scores matrix when variable in the last column can be right-censored and the other variables are mixed continuous/discrete
}
